# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API for Ranja Andriamiadana
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface SetupAdminRequest {
  username: string;
  password: string;
  email?: string;
}

export interface AdminUser {
  id: number;
  username: string;
  email?: string;
}

export interface LoginResponse {
  token: string;
  user: AdminUser;
}

export interface NeedsSetupResponse {
  needsSetup: boolean;
}

export type ProjectMediaType =
  (typeof ProjectMediaType)[keyof typeof ProjectMediaType];

export const ProjectMediaType = {
  image: "image",
  video: "video",
  youtube: "youtube",
} as const;

export interface ProjectMedia {
  id: number;
  projectId: number;
  type: ProjectMediaType;
  url: string;
  caption?: string;
  order?: number;
}

export interface Project {
  id: number;
  title: string;
  description: string;
  longDescription?: string;
  category: string;
  year: number;
  technologies: string[];
  githubUrl?: string;
  demoUrl?: string;
  featured: boolean;
  color?: string;
  icon?: string;
  coverImage?: string;
  media?: ProjectMedia[];
  createdAt?: string;
}

export interface CreateProjectRequest {
  title: string;
  description: string;
  longDescription?: string;
  category: string;
  year: number;
  technologies: string[];
  githubUrl?: string;
  demoUrl?: string;
  featured?: boolean;
  color?: string;
  icon?: string;
  coverImage?: string;
}

export type AddMediaRequestType =
  (typeof AddMediaRequestType)[keyof typeof AddMediaRequestType];

export const AddMediaRequestType = {
  image: "image",
  video: "video",
  youtube: "youtube",
} as const;

export interface AddMediaRequest {
  type: AddMediaRequestType;
  url: string;
  caption?: string;
  order?: number;
}

export interface Profile {
  id: number;
  name: string;
  title: string;
  bio?: string;
  email?: string;
  phone?: string;
  location?: string;
  github?: string;
  linkedin?: string;
  avatarUrl?: string;
}

export interface UpdateProfileRequest {
  name?: string;
  title?: string;
  bio?: string;
  email?: string;
  phone?: string;
  location?: string;
  github?: string;
  linkedin?: string;
  avatarUrl?: string;
}

export interface Skill {
  id: number;
  name: string;
  category: string;
  level: number;
  icon?: string;
  color?: string;
}

export interface CreateSkillRequest {
  name: string;
  category: string;
  level: number;
  icon?: string;
  color?: string;
}

export type GetProjectsParams = {
  category?: string;
  featured?: boolean;
};