# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Portfolio API for Ranja Andriamiadana
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: projects
    description: Portfolio projects and exploits
  - name: skills
    description: Skills and technologies
  - name: profile
    description: Profile management

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/setup:
    post:
      operationId: setupAdmin
      tags: [auth]
      summary: Create first admin if none exists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupAdminRequest"
      responses:
        "201":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "409":
          description: Admin already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/needs-setup:
    get:
      operationId: checkNeedsSetup
      tags: [auth]
      summary: Check if initial admin setup is needed
      responses:
        "200":
          description: Setup status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NeedsSetupResponse"

  /projects:
    get:
      operationId: getProjects
      tags: [projects]
      summary: Get all projects (public)
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /projects/{id}/media:
    post:
      operationId: addProjectMedia
      tags: [projects]
      summary: Add media URL to a project (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMediaRequest"
      responses:
        "201":
          description: Media added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMedia"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get portfolio owner profile (public)
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  /skills:
    get:
      operationId: getSkills
      tags: [skills]
      summary: Get all skills (public)
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"

    post:
      operationId: createSkill
      tags: [skills]
      summary: Create a skill (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSkillRequest"
      responses:
        "201":
          description: Skill created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"

  /skills/{id}:
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Delete a skill (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    SetupAdminRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
      required:
        - token
        - user

    NeedsSetupResponse:
      type: object
      properties:
        needsSetup:
          type: boolean
      required:
        - needsSetup

    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
      required:
        - id
        - username

    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        longDescription:
          type: string
        category:
          type: string
        year:
          type: integer
        technologies:
          type: array
          items:
            type: string
        githubUrl:
          type: string
        demoUrl:
          type: string
        featured:
          type: boolean
        color:
          type: string
        icon:
          type: string
        coverImage:
          type: string
        media:
          type: array
          items:
            $ref: "#/components/schemas/ProjectMedia"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - category
        - year
        - technologies
        - featured

    CreateProjectRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        longDescription:
          type: string
        category:
          type: string
        year:
          type: integer
        technologies:
          type: array
          items:
            type: string
        githubUrl:
          type: string
        demoUrl:
          type: string
        featured:
          type: boolean
        color:
          type: string
        icon:
          type: string
        coverImage:
          type: string
      required:
        - title
        - description
        - category
        - year
        - technologies

    ProjectMedia:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
          enum: [image, video, youtube]
        url:
          type: string
        caption:
          type: string
        order:
          type: integer
      required:
        - id
        - projectId
        - type
        - url

    AddMediaRequest:
      type: object
      properties:
        type:
          type: string
          enum: [image, video, youtube]
        url:
          type: string
        caption:
          type: string
        order:
          type: integer
      required:
        - type
        - url

    Profile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        bio:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        github:
          type: string
        linkedin:
          type: string
        avatarUrl:
          type: string
      required:
        - id
        - name
        - title

    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        bio:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        github:
          type: string
        linkedin:
          type: string
        avatarUrl:
          type: string

    Skill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        level:
          type: integer
        icon:
          type: string
        color:
          type: string
      required:
        - id
        - name
        - category
        - level

    CreateSkillRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        level:
          type: integer
        icon:
          type: string
        color:
          type: string
      required:
        - name
        - category
        - level