# lib/api-zod/src/generated/types/addMediaRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API for Ranja Andriamiadana
 * OpenAPI spec version: 0.1.0
 */
import type { AddMediaRequestType } from "./addMediaRequestType";

export interface AddMediaRequest {
  type: AddMediaRequestType;
  url: string;
  caption?: string;
  order?: number;
}