# lib/api-zod/src/generated/types/needsSetupResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API for Ranja Andriamiadana
 * OpenAPI spec version: 0.1.0
 */

export interface NeedsSetupResponse {
  needsSetup: boolean;
}