# lib/api-zod/src/generated/types/projectMedia.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API for Ranja Andriamiadana
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectMediaType } from "./projectMediaType";

export interface ProjectMedia {
  id: number;
  projectId: number;
  type: ProjectMediaType;
  url: string;
  caption?: string;
  order?: number;
}