# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API for Ranja Andriamiadana
 * OpenAPI spec version: 0.1.0
 */

export interface Skill {
  id: number;
  name: string;
  category: string;
  level: number;
  icon?: string;
  color?: string;
}